download docker-compose on demand in ComposeCmd
