+++ pkg/ddevapp/ddevapp.go
@@ -826,10 +826,6 @@ func (app *DdevApp) Start() error {
 	if err != nil {
 		return err
 	}
-	err = DownloadDockerComposeIfNeeded(app)
-	if err != nil {
-		return err
-	}
 
 	err = app.ProcessHooks("pre-start")
 	if err != nil {
